@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Devices
@@ -19,12 +18,12 @@ namespace Microsoft.Azure.Devices
         /// Results of the metrics collection queries
         /// </summary>
         [JsonProperty("results")]
-        public IDictionary<string, long> Results { get; set; } = new Dictionary<string, long>();
+        public IDictionary<string, long> Results { get; } = new Dictionary<string, long>();
 
         /// <summary>
         /// Queries used for metrics collection
         /// </summary>
         [JsonProperty("queries")]
-        public IDictionary<string, string> Queries { get; set; } = new Dictionary<string, string>();
+        public IDictionary<string, string> Queries { get; } = new Dictionary<string, string>();
     }
 }
