@@ -294,7 +294,7 @@ namespace Microsoft.Azure.Devices.Provisioning.Service
         /// The list of names of IoT hubs the device in this resource can be allocated to. Must be a subset of tenant level list of IoT hubs
         /// </summary>
         [JsonProperty(PropertyName = "iotHubs", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public ICollection<string> IotHubs { get; set; }
+        public IList<string> IotHubs { get; set; } = new List<string>();
 
         /// <summary>
         /// Convert this object in a pretty print format.
